requirements.txt
python-telegram-bot==21.0.1
python-dotenv==1.0.1
asyncpg==0.29.0